tests/wpt/tests/css/css-masonry/tentative/parsing/masonry-fill-computed.html

<!DOCTYPE html>
<html>
<head>
  <meta charset="utf-8">
  <title>CSS Masonry: masonry-fill getComputedStyle()</title>
  <link rel="author" title="Celeste Pan" href="mailto:celestepan@microsoft.com">
  <link rel="help" href="https://drafts.csswg.org/css-grid-3">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/css/support/computed-testcommon.js"></script>
  <script src="/css/support/inheritance-testcommon.js"></script>
</head>
<body>
  <div id="target"></div>
  </div>
  <script>
    test_computed_value("masonry-fill", "normal");
    test_computed_value("masonry-fill", "reverse");
  </script>
</body>
</html>
